VIQ-2005 Update keyword views to use es components instead of sdb

+++ utils/es_components_api_utils.py
@@ -69,9 +69,9 @@ class QueryGenerator:
 
                 query = QueryBuilder().build().must().range().field(field)
                 if min:
-                    query = query.gte(int(min))
+                    query = query.gte(float(min))
                 if max:
-                    query = query.lte(int(max))
+                    query = query.lte(float(max))
                 filters.append(query.get())
 
         return filters
